+++ src/components/Form.tsx
@@ -1,0 +1,11 @@
+import { styled } from 'styled-components';
+export const StForm = styled.form`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  margin-top: 15%;
+  width: 70%;
+  background-color: white;
+  border-radius: 8px;
+`;
