+++ src/components/Logo.tsx
@@ -1,0 +1,15 @@
+import { styled } from 'styled-components';
+import { ReactComponent as Logo } from '@imgs/start/logo-graphic.svg';
+
+export const MainLogo = styled(Logo)`
+  fill: var(--color-main);
+  width: 103px;
+  height: 116px;
+`;
+
+export const BulletLogo = styled(Logo)`
+  width: 39.27px;
+  height: 64.4px;
+  margin-top: 50px;
+  fill: var(--color-main);
+`;
